requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argilla==1.8.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.1
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
dataclasses-json==0.5.7
Deprecated==1.2.14
distlib==0.3.4
Django==4.0.4
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-shopping-cart==0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.6
duckdb==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.96.0
filelock==3.6.0
flatbuffers==23.5.26
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.194
langchainplus-sdk==0.0.7
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
olefile==0.46
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.5.4.58
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.1.0
platformdirs==2.5.2
posthog==3.0.1
protobuf==4.23.2
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.4.0
pypandoc==1.11
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2022.1
PyYAML==6.0
razorpay==1.3.0
regex==2023.6.3
requests==2.28.1
rfc3986==1.5.0
rich==13.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
sqlparse==0.4.2
starlette==0.27.0
swagger-spec-validator==3.0.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2022.1
unstructured==0.7.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.22.0
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
